+++ controller/server/view/network_list_page.ml
@@ -1,15 +1,20 @@
 open Connman.Service
 open Tyxml.Html
 
-let service { id; name; strength; ipv4 } =
+let service_item ({ id; name; strength; ipv4 } as service) =
   let strength =
     match strength with
     | Some s -> [ Signal_strength.html s ]
     | None -> []
   in
+  let
+    classes =
+      [ "d-NetworkList__Network" ]
+        @ if Connman.Service.is_connected service then [ "d-NetworkList__Network--Connected" ] else []
+  in
   li
     [ a
-        ~a:[ a_class [ "d-NetworkList__Network" ]
+        ~a:[ a_class classes
         ; a_href ("/network/" ^ id)
         ]
         [ div [ txt name ]
@@ -34,17 +39,30 @@ type params =
   }
 
 let html { proxy; is_internet_connected; services; interfaces } =
+  let connected_services, available_services =
+    List.partition Connman.Service.is_connected services
+  in
   Page.html ~current_page:Page.Network (
     div
-      [ h1 ~a:[ a_class [ "d-Title" ] ] [ txt "Network" ]
+      [ div ~a:[ a_class [ "d-Title" ] ]
+           [ div
+               ~a:[ a_class [ "d-Network__Title" ] ]
+               [ h1 [ txt "Network" ]
+               ; div
+                   ~a:[ a_class [ "d-Network__Refresh" ] ]
+                   [ a
+                       ~a:[ a_href "/network?timeout=3"
+                       ; a_class [ "d-Button" ]
+                       ]
+                       [ txt "Refresh" ]
+                   ]
+               ]
+           ]
 
-      ; div
-          ~a:[ a_class [ "d-Network__Refresh" ] ]
-          [ a
-              ~a:[ a_href "/network?timeout=3"
-              ; a_class [ "d-Button" ]
-              ]
-              [ txt "Refresh" ]
+      ; section
+          [ ul
+              ~a:[ a_class [ "d-NetworkList" ]; a_role [ "list" ] ]
+              (List.map service_item connected_services)
           ]
 
       ; Definition.list (
@@ -71,13 +89,13 @@ let html { proxy; is_internet_connected; services; interfaces } =
         )
 
       ; section
-          [ h2 ~a:[ a_class [ "d-Subtitle" ] ] [ txt "Services" ]
-          ; if List.length services = 0 then
-              txt "No services available"
+          [ h2 ~a:[ a_class [ "d-Subtitle" ] ] [ txt "Available Networks" ]
+          ; if List.length available_services = 0 then
+              txt "No networks available"
             else
               ul
                 ~a:[ a_class [ "d-NetworkList" ]; a_role [ "list" ] ]
-                (List.map service services)
+                (List.map service_item available_services)
           ]
 
       ; section